Refactor GhDocumentTool into partial classes
Split the 793-line GhDocumentTool.cs into 2 files for better
maintainability and code navigation:

- GhDocumentTool.cs: Core dispatch, ToolInfo, document/solver/snapshot operations
- GhDocumentTool.Capture.cs: Canvas and viewport capture actions with helper methods

No functional changes - the MCP tool API remains identical.
